# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards management system
  description: API Description for Bank Cards management system
  contact:
    name: Alex VM
    email: my@email.com
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/cards/transfer:
    post:
      tags:
      - Cards
      summary: Операция перевода Юзером средств между своими картами
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardTransferDto"
        required: true
      responses:
        "200":
          description: Операция успешно завершена
          content:
            text/plain:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/cards/admin/create:
    post:
      tags:
      - Cards
      summary: Регистрация Админом новой Карты для Юзера
      description: |
        Номер карты валидируется по алгоритму Луна. Чтобы получить валидный номер (при тестировании),
        можно сгенерировать его с помощью эндпойнта 'cards/admin/generate-card-number'. Срок действия
        новой карты должен быть не ранее текущей даты (карта не должна быть "просрочена").
        Баланс карты должен числом СТРОГО С ДВУМЯ ЗНАКАМИ после запятой.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateDto"
        required: true
      responses:
        "201":
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardRespDto"
        "400":
          $ref: "#/components/responses/Validation"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
...
  /api/v1/cards/block/{cardId}:
    patch:
      tags:
      - Cards
      summary: Юзер помечает карту для блокировки Админом
      operationId: blockRequest
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Операция успешно завершена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardRespDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/cards/admin/change-status/{cardId}:
    patch:
      tags:
      - Cards
      summary: "Админ может изменить статус карты. Если текущий статус EXPIRED, то\
        \ изменить нельзя (получим исключение)."
      operationId: changeStatus
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardStatusDto"
        required: true
      responses:
        "200":
          description: Операция успешно завершена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardRespDto"
        "400":
          $ref: "#/components/responses/Validation"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
...
  /api/v1/cards/my-cards:
    get:
      tags:
      - Cards
      summary: Получение Юзером списка принадлежащих ему карт (с фильтрацией и пагинацией)
      description: При отсутствии параметров фильтрации возвращается список ВСЕХ карт.
      operationId: getAllForMe
      parameters:
      - name: dto
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/CardHolderListDto"
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "id,ASC"
          items:
            type: string
      responses:
        "200":
          description: Page of cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardPageResponse"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/cards/my-balance:
    get:
      tags:
      - Cards
      summary: Получить сумму балансов всех карт текущего пользователя
      description: Возвращает сумму балансов всех карт пользователя
      operationId: getTotalBalance
      responses:
        "200":
          description: Сумма успешно получена
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/cards/admin/list:
    get:
      tags:
      - Cards
      summary: Получение Админом списка всех карт (с фильтрацией и пагинацией)
      description: При отсутствии параметров фильтрации возвращается список ВСЕХ карт.
      operationId: getCardsList
      parameters:
      - name: dto
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/CardsListDto"
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "id,ASC"
          items:
            type: string
      responses:
        "200":
          description: Page of cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardPageResponse"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/cards/admin/generate-card-number:
    get:
      tags:
      - Cards
      summary: Генерация номера кредитной карты
      description: Для целей тестирования. Генерирует валидный номер банковской карты
        (16 цифр) с контрольной по алгоритму Луна.
      operationId: generate
      responses:
        "200":
          description: Card Number
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
  /api/v1/cards/admin/delete/{cardId}:
    delete:
      tags:
      - Cards
      summary: Админ может удалить BLOCKED или EXPIRED карту
      description: Удаляет карту и возвращает 204 No Content
      operationId: delete
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Карта успешно удалена
          content:
            '*/*':
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseSchema"
      security:
      - JWT Bearer: []
components:
  schemas:
    CardTransferDto:
      type: object
      properties:
        fromId:
          type: integer
          format: int64
          example: 2
          minimum: 1
        toId:
          type: integer
          format: int64
          example: 1
          minimum: 1
        amount:
          type: number
          example: 1234.56
          minimum: 0.01
      required:
      - amount
      - fromId
      - toId
    ErrorResponseSchema:
      type: object
      description: Unified error response
      properties:
        type:
          type: string
          description: "The string containing a URI reference [URI] that identifies\
            \ the problem type."
          example: about:blank
        title:
          type: string
          example: VALIDATION_ERROR
        status:
          type: integer
          format: int32
          description: The number indicating the HTTP status code
          example: 400
        detail:
          type: string
          example: Validation failed
        instance:
          type: string
          description: The string containing a URI reference that identifies the specific
            occurrence of the problem
          example: /users/admin
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors (only for VALIDATION_ERROR)
          example:
            email: must be a valid email
    CardCreateDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 51
          minimum: 1
        cardNumber:
          type: string
          example: 4000006806224829
        expiryDate:
          type: string
          pattern: "^(0[1-9]|1[0-2])/\\d{2}$"
        balance:
          type: number
          example: 1234.55
      required:
      - balance
      - cardNumber
      - expiryDate
      - userId
    CardRespDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardHolder:
          type: string
        cardNumber:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        toBlock:
          type: boolean
        balance:
          type: number
        userId:
          type: integer
          format: int64
...
    CardStatusDto:
      type: object
      properties:
        newStatus:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
...
    PageMeta:
      type: object
      description: Pagination metadata
      properties:
        size:
          type: integer
          format: int32
          example: 10
        number:
          type: integer
          format: int32
          example: 0
        totalElements:
          type: integer
          format: int64
          example: 2
        totalPages:
          type: integer
          format: int32
          example: 1
...
    CardHolderListDto:
      type: object
      properties:
        status:
          type: string
          description: Card status
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          example: ACTIVE/BLOCKED/EXPIRED
        toBlock:
          type: boolean
          description: User's request to block the Card
          example: false
        expiryDate:
          type: string
          description: Card expiry date MM/yy
          example: 12/25
          pattern: "^(0[1-9]|1[0-2])/\\d{2}$"
    CardPageResponse:
      type: object
      properties:
        content:
          type: array
          description: Page content
          items:
            $ref: "#/components/schemas/CardRespDto"
        page:
          $ref: "#/components/schemas/PageMeta"
          description: Pagination info
    CardsListDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          minimum: 1
        cardHolder:
          type: string
          pattern: "[A-Za-z]+"
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        toBlock:
          type: boolean
        expiryDate:
          type: string
          pattern: "^(0[1-9]|1[0-2])/\\d{2}$"
  responses:
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
    Validation:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
  securitySchemes:
    JWT Bearer:
      type: http
      description: JWT in header
      in: header
      scheme: bearer
      bearerFormat: JWT